requirements.txt
Django==1.8.2
django-appconf==0.5
django-eztables==0.3.2
-e git+https://github.com/shahidgadda/django-jfu.git@f45025f0a6d4cbc8d7f4933a77da8468cd296159#egg=django_jfu-master
django-multiupload==0.5
django-user-accounts==1.1
django.js==0.8.1
MySQL-python==1.2.5
Pillow==2.8.2
pinax-theme-bootstrap-account==1.0b2
pytz==2015.4
six==1.9.0